# static/openapi.yaml
openapi: 3.0.3
info:
  title: CarrierShark API v1
  version: 1.0.0
  description: |
    The CarrierShark API allows companies to monitor carriers, receive compliance alerts, and manage digital carrier contracts.

    Base URL:
    https://carriershark.com/api/v1

    Authentication:
    All endpoints require a company API key.

    Authorization header format:

    Authorization: Bearer cs_live_your_api_key

    Test key example:
    cs_test_123456789
  contact:
    name: CarrierShark Support
    email: support@carriershark.com
    url: https://carriershark.com
  license:
    name: Proprietary
servers:
  - url: https://carriershark.com/api/v1
    description: Production v1 API
security:
  - bearerAuth: []
...
paths:
...
  /carriers:
    get:
      tags: [Carriers]
      summary: Search carrier database
      description: |
        Lookup-oriented carrier search endpoint for focused carrier retrieval.
        Broader extraction workflows should use saved carriers or enterprise workflows.
        Strict rate limits are intentional to discourage bulk polling.

        Allowed searches:
        - dot
        - mc
        - legalname with 3+ characters
        - dbaname with 3+ characters
        - city + state

        Blocked searches:
        - state alone
        - city alone
        - legalname shorter than 3 chars
        - dbaname shorter than 3 chars
      operationId: searchCarriers
      parameters:
        - $ref: '#/components/parameters/DotQuery'
        - name: mc
          in: query
          schema:
            type: string
        - name: legalname
          in: query
          schema:
            type: string
        - name: dbaname
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
        - name: page
          in: query
          description: Page number of results (starting at 1, maximum 5 for this endpoint).
          schema:
            type: integer
            minimum: 1
            maximum: 5
            default: 1
        - name: pageSize
          in: query
          description: Number of results per page (maximum 25 for this endpoint).
          schema:
            type: integer
            minimum: 1
            maximum: 25
            default: 25
      responses:
        '200':
          description: Carrier search results.
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarrierListResponse'
              example:
                rows:
                  - dot: "336075"
                    dotnumber: "336075"
                    legalname: ACME TRANSPORT LLC
                    dbaname: ACME LOGISTICS
                    mc_number: "MC-654321"
                    phystreet: 123 MAIN ST
                    phycity: DALLAS
                    phystate: TX
                    phyzipcode: "75201"
                    phycountry: US
                    allowedtooperate: AUTHORIZED
                    safetyrating: SATISFACTORY
                    statuscode: A
                    totaldrivers: 12
                    totalpowerunits: 10
                    retrieval_date: '2026-03-15T00:00:00.000Z'
                    last_polled_at: '2026-03-30T18:42:31.000Z'
                    insurance_summary:
                      has_on_file: true
                      has_structured_coverages: true
                      is_expired: false
                      next_expiration_date: '2026-12-31'
                      auto_liability_limit: 1000000
                      cargo_limit: 100000
                      general_liability_limit: 1000000
                      coverage_count: 3
                total: 1
                page: 1
                pageSize: 25
        '400':
          description: Invalid carrier search input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                requiresValidSearch:
                  value:
                    error: Search requires dot, mc, legalname (3+ chars), dbaname (3+ chars), or city + state.
                stateAlone:
                  value:
                    error: state alone is too broad; combine it with city, legalname, or dbaname.
                cityAlone:
                  value:
                    error: city alone is too broad; include state.
                pageSizeLimit:
                  value:
                    error: pageSize cannot exceed 25.
                pageLimit:
                  value:
                    error: page cannot exceed 5 for carrier search.
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Carrier search rate limit exceeded. Please slow down.
        '500':
          description: Search failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Search failed
  /carriers/{dot}:
    get:
      tags: [Carriers]
      summary: Get carrier
      description: |
        Returns a single carrier record for the provided USDOT number.
        Includes standard carrier fields plus `cargo_carried` when available.
      operationId: getCarrierByDot
      parameters:
        - $ref: '#/components/parameters/DotPath'
      responses:
        '200':
          description: Carrier record with cargo information.
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarrierDetail'
              example:
                dot: "336075"
                dotnumber: "336075"
                legalname: ACME TRANSPORT LLC
                dbaname: ACME LOGISTICS
                mc_number: "MC-654321"
                phystreet: 123 MAIN ST
                phycity: DALLAS
                phystate: TX
                phyzipcode: "75201"
                phycountry: US
                allowedtooperate: AUTHORIZED
                safetyrating: SATISFACTORY
                statuscode: A
                totaldrivers: 12
                totalpowerunits: 10
                retrieval_date: '2026-03-15T00:00:00.000Z'
                last_polled_at: '2026-03-30T18:42:31.000Z'
                address1: 123 MAIN ST
                address2: null
                city: DALLAS
                state: TX
                zip: "75201"
                retrieval_date_formatted: Mar 15, 2026 12:00 AM EST
                cargo_carried:
                  - General Freight
                  - Machinery
                insurance_coverages:
                  - id: 981
                    document_id: 442
                    coverage_type: AUTO LIABILITY
                    insurer_name: Great West Casualty Company
                    policy_number: GL-4459011
                    effective_date: '2026-01-01'
                    expiration_date: '2026-12-31'
                    limit_amount: 1000000
                    status: ACTIVE
                    is_expired: false
                  - id: 982
                    document_id: 442
                    coverage_type: CARGO
                    insurer_name: Great West Casualty Company
                    policy_number: CG-2087710
                    effective_date: '2026-01-01'
                    expiration_date: '2026-12-31'
                    limit_amount: 100000
                    status: ACTIVE
                    is_expired: false
                insurance_summary:
                  has_on_file: true
                  has_structured_coverages: true
                  is_expired: false
                  next_expiration_date: '2026-12-31'
                  auto_liability_limit: 1000000
                  cargo_limit: 100000
                  general_liability_limit: 1000000
                  coverage_count: 2
        '404':
          description: Carrier not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Carrier not found
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Rate limit exceeded
        '500':
          description: Database query failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Database query failed
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API_KEY
      description: |
        Authenticate using your CarrierShark company API key.

        Example:

        Authorization: Bearer cs_live_your_api_key_here
  headers:
    X-RateLimit-Limit:
      schema:
        type: integer
      description: Maximum number of requests allowed in the current window.
    X-RateLimit-Remaining:
      schema:
        type: integer
      description: Requests remaining in the current rate limit window.
    X-RateLimit-Reset:
      schema:
        type: integer
      description: UNIX timestamp when the rate limit window resets.
  parameters:
    DotPath:
      name: dot
      in: path
      required: true
      schema:
        type: string
        pattern: '^[0-9]+$'
        example: "336075"
      description: USDOT number.
    DotQuery:
      name: dot
      in: query
      schema:
        type: string
        pattern: '^[0-9]+$'
        example: "336075"
      description: USDOT number.
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    InsuranceSummary:
      type: object
      required:
        - has_on_file
        - has_structured_coverages
        - is_expired
        - next_expiration_date
        - auto_liability_limit
        - cargo_limit
        - general_liability_limit
        - coverage_count
      properties:
        has_on_file:
          type: boolean
        has_structured_coverages:
          type: boolean
        is_expired:
          type: boolean
        next_expiration_date:
          type: string
          format: date
          nullable: true
        auto_liability_limit:
          type: number
          nullable: true
        cargo_limit:
          type: number
          nullable: true
        general_liability_limit:
          type: number
          nullable: true
        coverage_count:
          type: integer
    InsuranceCoverage:
      type: object
      properties:
        id:
          oneOf:
            - type: integer
            - type: string
              format: uuid
          nullable: true
        document_id:
          oneOf:
            - type: integer
            - type: string
              format: uuid
          nullable: true
        coverage_type:
          type: string
          enum:
            - AUTO LIABILITY
            - CARGO
            - GENERAL LIABILITY
            - WORKERS COMP
            - OTHER
        insurer_name:
          type: string
          nullable: true
        policy_number:
          type: string
          nullable: true
        effective_date:
          type: string
          format: date
          nullable: true
        expiration_date:
          type: string
          format: date
          nullable: true
        limit_amount:
          type: number
          nullable: true
        status:
          type: string
          enum: [ACTIVE, EXPIRED]
        is_expired:
          type: boolean
    Carrier:
      type: object
      description: |
        Carrier row selected from `carriers` as `c.*` plus a compatibility alias `dot`.
        Fields listed below are the stable/common fields documented by the API.
      properties:
        dot:
          type: string
        dotnumber:
          type: string
        legalname:
          type: string
          nullable: true
        dbaname:
          type: string
          nullable: true
        mc_number:
          type: string
          nullable: true
        phystreet:
          type: string
          nullable: true
        phycity:
          type: string
          nullable: true
        phystate:
          type: string
          nullable: true
        phyzipcode:
          type: string
          nullable: true
        phycountry:
          type: string
          nullable: true
        allowedtooperate:
          type: string
          nullable: true
        safetyrating:
          type: string
          nullable: true
        statuscode:
          type: string
          nullable: true
        totaldrivers:
          type: integer
          nullable: true
        totalpowerunits:
          type: integer
          nullable: true
        retrieval_date:
          type: string
          format: date-time
          nullable: true
        last_polled_at:
          type: string
          format: date-time
          nullable: true
        insurance_summary:
          $ref: '#/components/schemas/InsuranceSummary'
      additionalProperties: true
    CarrierListResponse:
      type: object
      required: [rows, total, page, pageSize]
      properties:
        rows:
          type: array
          items:
            $ref: '#/components/schemas/Carrier'
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
...
    CarrierDetail:
      type: object
      description: Carrier detail row from `/carriers/{dot}` including derived cargo and compatibility fields.
      allOf:
        - $ref: '#/components/schemas/Carrier'
        - type: object
          properties:
            address1:
              type: string
              nullable: true
            address2:
              type: string
              nullable: true
            city:
              type: string
              nullable: true
            state:
              type: string
              nullable: true
            zip:
              type: string
              nullable: true
            retrieval_date_formatted:
              type: string
              nullable: true
            cargo_carried:
              type: array
              items:
                type: string
            insurance_summary:
              $ref: '#/components/schemas/InsuranceSummary'
            insurance_coverages:
              type: array
              items:
                $ref: '#/components/schemas/InsuranceCoverage'
          additionalProperties: true
...